# setup.sh: Decrease setup time on Fedora