The command to check if a package is installed on a Fedora system take 10x the time of Arch and Ubuntu counterparts. This makes the 'setup.sh -i' take 4x slower on Fedora. The measure time assumes that all the all packages are already installed, because it would add unneeded variables.

As an example, my computer running takes 2.5s on average to run `setup.sh -i`. On Fedora, 11.8s. After the change Fedora takes 3s.

Some raw backing data:
```
# BEFORE CHANGE
[psato@fedora kworkflow-gitrepo]$ hyperfine './setup.sh -i'
Benchmark 1: ./setup.sh -i
  Time (mean ± σ):     11.896 s ±  0.182 s    [User: 9.642 s, System: 1.911 s]
  Range (min … max):   11.706 s … 12.376 s    10 runs

#AFTER CHANGE
[psato@fedora kworkflow-gitrepo]$ hyperfine './setup.sh -i'
Benchmark 1: ./setup.sh -i
  Time (mean ± σ):      3.006 s ±  0.379 s    [User: 2.245 s, System: 0.440 s]
  Range (min … max):    2.496 s …  3.572 s    10 runs
```